フロアを分割
