+++ client/components/Floor/Floor.module.css
@@ -1,0 +1,133 @@
+.card {
+  margin-bottom: 1rem;
+  background-color: white;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgb(0 0 0 / 10%);
+}
+
+.cardTitle {
+  display: flex;
+  gap: 0.5rem;
+  align-items: center;
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: #1e40af;
+}
+
+.icon {
+  width: 1.125rem;
+  height: 1.125rem;
+}
+
+.button {
+  display: block;
+  width: 100%;
+  padding: 0.75rem;
+  margin-top: 0.75rem;
+  font-size: 1rem;
+  color: white;
+  cursor: pointer;
+  background-color: #1e40af;
+  border: none;
+  border-radius: 4px;
+  transition: background-color 0.3s ease;
+}
+
+.button:hover,
+.button:focus {
+  background-color: #1e3a8a;
+}
+
+.facilityButton {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: auto;
+  aspect-ratio: 1 / 1;
+  padding: 0.75rem;
+  font-size: 0.75rem;
+  cursor: pointer;
+  background-color: #e2e8f0;
+  border: none;
+  border-radius: 8px;
+  transition: background-color 0.3s ease;
+}
+
+.facilityButton:hover,
+.facilityButton:focus {
+  background-color: #cbd5e1;
+}
+
+.facilityIcon {
+  width: 1.5rem;
+  height: 1.5rem;
+  margin-bottom: 0.25rem;
+}
+
+.facilityName {
+  text-align: center;
+  word-break: break-word;
+}
+
+.note {
+  margin-top: 0.75rem;
+  font-size: 0.75rem;
+  color: #64748b;
+}
+
+.floorSelection,
+.isActive {
+  display: flex;
+  justify-content: space-around;
+  margin-bottom: 20px;
+}
+
+.floorButton {
+  padding: 10px 20px;
+  cursor: pointer;
+  background-color: #f0f0f0;
+  border: 1px solid #ddd;
+  border-radius: 5px;
+}
+
+.floorButton:hover {
+  background-color: #e0e0e0;
+}
+
+.activeFloor {
+  color: white;
+  background-color: #0070f3;
+}
+
+.notActive {
+  display: none;
+}
+
+@media (width >= 640px) {
+  .main {
+    max-width: 768px;
+    padding: 2rem;
+    margin: 0 auto;
+  }
+
+  .card {
+    margin-bottom: 2rem;
+  }
+
+  .cardTitle {
+    font-size: 1.25rem;
+  }
+
+  .facilityGrid {
+    grid-template-columns: repeat(3, 1fr);
+  }
+
+  .facilityButton {
+    font-size: 0.875rem;
+  }
+
+  .note {
+    font-size: 0.875rem;
+  }
+}
